fixed old textures, added new models
i spent the whole day pretty much separating that stage model and creating new texture for the assets , it might not of been needed but you can see the cracks in the texture on the old speaker on the left beside the new texture on the right. i fixed this on all the models except the main character, will finish that wednesday hopefully. also textured some of keiths models that i diddnt get around to before. i also created a simple drink coaster if you want to try implement the trow mechanic on it at some point